Trim trailing slash
